--- src/main/java/com/leoh/bloomit/board/entity/Board.java
@@ -1,24 +1,0 @@
-package com.leoh.bloomit.board.entity;
-
-import com.leoh.bloomit.boardtype.entity.BoardType;
-import com.leoh.bloomit.common.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Board extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private BoardType boardType;
-
-    private Long relatedId;
-
-}
