--- src/main/java/com/leoh/bloomit/book/entity/Book.java
@@ -1,42 +1,0 @@
-package com.leoh.bloomit.book.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Book {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String isbn;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    private String author;
-
-    @Column(nullable = false)
-    private String publisher;
-
-    @Column(nullable = false)
-    private LocalDateTime publicationDate;
-
-    @Lob
-    private String imageUrl;
-
-    @Lob
-    private String story;
-
-    private String description;
-
-}
