--- src/main/java/com/leoh/bloomit/bookreport/entity/BookReport.java
@@ -1,30 +1,0 @@
-package com.leoh.bloomit.bookreport.entity;
-
-import com.leoh.bloomit.book.entity.Book;
-import com.leoh.bloomit.common.entity.BaseEntity;
-import com.leoh.bloomit.member.entity.Member;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class BookReport extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "book_id")
-    private Book book;
-
-    private int viewCount;
-
-}
