+++ src/main/java/com/leoh/bloomit/domain/board/entity/Board.java
@@ -1,0 +1,25 @@
+package com.leoh.bloomit.domain.board.entity;
+
+import com.leoh.bloomit.domain.boardtype.entity.BoardType;
+import com.leoh.bloomit.common.entity.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Board extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "board_type_id")
+    private BoardType boardType;
+
+    private Long relatedId;
+
+}
