+++ src/main/java/com/leoh/bloomit/domain/comment/entity/Comment.java
@@ -1,0 +1,32 @@
+package com.leoh.bloomit.domain.comment.entity;
+
+import com.leoh.bloomit.common.entity.BaseEntity;
+import com.leoh.bloomit.domain.member.entity.Member;
+import com.leoh.bloomit.domain.post.entity.Post;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Comment extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id", nullable = false)
+    private Post post;
+
+    @Lob
+    @Column(nullable = false)
+    private String content;
+
+}
