+++ src/main/java/com/leoh/bloomit/domain/like/entity/Like.java
@@ -1,0 +1,29 @@
+package com.leoh.bloomit.domain.like.entity;
+
+import com.leoh.bloomit.domain.member.entity.Member;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Like {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    @Column(nullable = false)
+    private Long targetId;
+
+    @Column(nullable = false)
+    private String targetType;
+
+}
