+++ src/main/java/com/leoh/bloomit/domain/review/entity/Review.java
@@ -1,0 +1,35 @@
+package com.leoh.bloomit.domain.review.entity;
+
+import com.leoh.bloomit.domain.book.entity.Book;
+import com.leoh.bloomit.domain.member.entity.Member;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Review {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "book_id", nullable = false)
+    private Book book;
+
+    @Column(nullable = false, precision = 2, scale = 1)
+    private BigDecimal rating;
+
+    @Lob
+    private String content;
+
+}
