--- src/main/java/com/leoh/bloomit/library/entity/Library.java
@@ -1,29 +1,0 @@
-package com.leoh.bloomit.library.entity;
-
-import com.leoh.bloomit.librarybook.entity.LibraryBook;
-import com.leoh.bloomit.member.entity.Member;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Library {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    @OneToOne(fetch = FetchType.LAZY)
-    private Member member;
-
-    @OneToMany(mappedBy = "library")
-    private List<LibraryBook> libraryBooks = new ArrayList<>();
-
-}
