--- src/main/java/com/leoh/bloomit/member/repository/MemberRepository.java
@@ -1,15 +1,0 @@
-package com.leoh.bloomit.member.repository;
-
-import com.leoh.bloomit.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByUsername(String username);
-
-    boolean existsByUsername(String username);
-
-}
